Remove unused Nix code
The initiative to move our build system to Nix was never completed, and
now it's certain it never will be. The unfinished work can sometimes
make it so that making a minor change or just switching branches can
cause everything to rebuild, which can take a hugely disruptive amount
of time. That's a high cost to pay for something we don't use, so we get
rid of it.